Move dimension_filters.py
